Fix Purescript `jQuery` dependency
`purescript-jquery` doesn't `require` `jquery` like other PureScript FFI users.
We used to get away with a global `jQuery` because we used it in custom
elements. But recent changes mean that we no longer rely on `jQuery` in custom
elements. Our solution is to create a foreign module for each application to
`require('jquery')`.